ml/classroom_people_counter/src/utils/config.py
import yaml
from pathlib import Path

def load_config(path='config/app_config.yaml'):
    p = Path(path)
    if not p.exists():
        raise FileNotFoundError(f'Config not found: {path}')
    with open(p, 'r', encoding='utf-8') as f:
        return yaml.safe_load(f)
