ml/classroom_people_counter/requirements.txt
opencv-python
numpy
pyyaml
deep_sort_realtime